Cambios en CategoryControllerTest

+++ Catalogo-Pruebas/src/main/java/com/example/domains/services/CategoryServiceImpl.java
@@ -18,8 +18,8 @@ import com.example.exceptions.InvalidDataException;
 import com.example.exceptions.NotFoundException;
 
 @Service
-public class CategoryServiceImpl implements CategoryService{
-	
+public class CategoryServiceImpl implements CategoryService {
+
 	@Autowired
 	private CategoryRepository dao;
 
@@ -30,7 +30,7 @@ public class CategoryServiceImpl implements CategoryService{
 
 	@Override
 	public <T> Iterable<T> getByProjection(Sort sort, Class<T> type) {
-		return dao.findAllBy(sort,type);
+		return dao.findAllBy(sort, type);
 	}
 
 	@Override
@@ -60,18 +60,15 @@ public class CategoryServiceImpl implements CategoryService{
 
 	@Override
 	public Category add(Category item) throws DuplicateKeyException, InvalidDataException {
-		if(item == null)
-		{
+		if (item == null) {
 			throw new InvalidDataException("[ERROR] El item no puede ser nulo");
 		}
-		
-		if(item.isInvalid())
-		{
+
+		if (item.isInvalid()) {
 			throw new InvalidDataException(item.getErrorsMessage());
 		}
-		
-		if(dao.existsById(item.getCategoryId()))
-		{
+
+		if (dao.existsById(item.getCategoryId())) {
 			throw new DuplicateKeyException(item.getErrorsMessage());
 		}
 		return dao.save(item);
@@ -79,18 +76,15 @@ public class CategoryServiceImpl implements CategoryService{
 
 	@Override
 	public Category modify(Category item) throws NotFoundException, InvalidDataException {
-		if(item == null)
-		{
+		if (item == null) {
 			throw new InvalidDataException("[ERROR] El item no puede ser nulo");
 		}
-		
-		if(item.isInvalid())
-		{
+
+		if (item.isInvalid()) {
 			throw new InvalidDataException(item.getErrorsMessage());
 		}
-		
-		if(!dao.existsById(item.getCategoryId()))
-		{
+
+		if (!dao.existsById(item.getCategoryId())) {
 			throw new NotFoundException();
 		}
 		return dao.save(item);
@@ -98,17 +92,23 @@ public class CategoryServiceImpl implements CategoryService{
 
 	@Override
 	public void delete(Category item) throws InvalidDataException {
-		if(item == null)
-		{
+		if (item == null) {
 			throw new InvalidDataException("[ERROR] El item no puede ser nulo");
 		}
-		
+
 		deleteById(item.getCategoryId());
 	}
 
 	@Override
-	public void deleteById(Integer id) {
-		dao.deleteById(id);
+	public void deleteById(Integer id) throws InvalidDataException {
+		try {
+			if (id < 0) {
+				throw new InvalidDataException("ERROR. La id no puede ser menor a 0");
+			}
+			dao.deleteById(id);
+		} catch (InvalidDataException e) {
+			throw new InvalidDataException(e.getMessage());
+		}
 	}
 
 	@Override
@@ -117,5 +117,4 @@ public class CategoryServiceImpl implements CategoryService{
 
 	}
 
-	
 }
